environmentd: move max_connections test to server.rs
To be with its friend.
